Add support to update project
